add user item

+++ app/Http/Controllers/ItemController.php
@@ -4,15 +4,23 @@ namespace App\Http\Controllers;
 
 use App\Models\Item;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class ItemController extends Controller
 {
     public function destroy($id)
-{
-    $item = Item::findOrFail($id);
-    $item->delete();
+    {
+        $item = Item::findOrFail($id);
+        
+        if ($item) {
 
-    return redirect()->route('home')->with('success', 'Item excluído com sucesso.');
-}
+            DB::table('user_item')->where('item_id', $id)->delete();
 
+            DB::table('item_genre')->where('item_id', $id)->delete();
+
+            $item->delete();
+
+            return redirect()->route('home');
+        }
+    }
 }
